Update index.yaml
I'm working on getting the Helm Charts added to Artifact Hub so anyone searching for the chart there will be able to find it. I added the repo, but am hitting this error: `error preparing package: error loading chart (https://helm.openwebui.com/open-webui-1.0.0.tgz): not found (package: open-webui version: 1.0.0)`

I believe this is because that version is still called out in `index.yaml`, but is not present in the `gh-pages` branch anymore. I'm hoping that removing it from the index file will allow Artifact Hub to pick up the valid versions of the chart correctly.